Langchain-Samples/FastApi_Apps/pyproject.toml
[project]
name = "fastapi-apps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "langchain-community>=0.3.21",
    "langchain-core>=0.3.54",
    "langchain-groq>=0.3.2",
    "langserve>=0.3.1",
    "python-dotenv>=1.1.0",
    "sse-starlette>=2.2.1",
    "streamlit>=1.44.1",
    "uvicorn>=0.34.2",
]
